fix(domain): car 디렉토리 변경, car 이동 상태 추가, move 결정 시 등록
Car.java CarTest.java 디렉토리 변경, MoveStatus.java 이동 상태 추가, move 결정 시 등록
 - Car.java CarTest.java 디렉토리 변경
   - Car.java, CarTest.java car 디렉토리 삭제 후 상위 racinggame 으로 이동
 - MoveStatus.java 이동 상태 추가
   - Car.java 에서 이동 상태를 기록하기 위해 movestatus 추가
   - isMovableForward 확인
   - 최소 이동 가능 횟수 MOVABLE_MINIMUM_NUMBER static 상수로 변경
