+++ src/test/java/racinggame/CarTest.java
@@ -1,4 +1,4 @@
-package racinggame.car;
+package racinggame;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
